Extract /about loader component and convert route to markdown post
