+++ src/lib/components/about/PersonCard.svelte
