+++ src/posts/about.md
@@ -1,0 +1,40 @@
+---
+title: About Us
+description: Staff behind PauseAI
+date: 2025-11-06
+---
+
+<script>
+	import AboutPeople from '$lib/components/about/AboutPeople.svelte'
+	import Link from '$lib/components/Link.svelte'
+</script>
+
+## How We Began
+
+We were founded in Utrecht, Netherlands in May 2023 by Joep Meindertsma, who put his job on hold because he couldn't ignore the existential risks from artificial intelligence any longer. We began with our first public action, which was a protest outside [Microsoft's Brussels lobbying office](/brussels-microsoft-protest).
+
+What started as one person's call to action has grown into a global grassroots movement with volunteers, [national leaders](/national-groups), and [local communities](/communities) across the world, all working toward the same goal: pausing frontier AI development until we can prove it's safe and keep it under democratic control.
+
+## Key Milestones
+
+- November 2023: Protested outside the inaugural [AI Safety Summit at Bletchley Park](/2023-november-uk).
+- February 2024: [Gathered outside OpenAI's headquarters in San Francisco](/2024-february)
+- May 2024: Organized coordinated protests across thirteen countries—including the US, UK, Brazil, Germany, Australia, and Norway ahead of the [AI Seoul Summit](/2024-may).
+- February 2025: Coordinated an event larger global protest for the [French AI Summit](/2025-february), adding Kenya and DRC to the mix.
+- June 2025: Led our [biggest protest yet outside Google DeepMind's London office](/deepmind-protest-2025), holding them accountable for breaking their safety commitments and hosted the first ever PauseCon training event.
+- Ongoing: Grew from a single-country initiative into a network with national organizations like PauseAI US, each running their own campaigns while we coordinate globally.
+
+## Non-violence
+
+We are committed to a peaceful approach to our advocacy.
+
+<AboutPeople />
+
+## Essential Information
+
+- [Privacy Policy](/privacy)
+- [Legal Info](/legal)
+- [Global Funding and Donors](/funding)
+- [Roadmap](/roadmap)
+- [Press](/press)
+- [Vacancies](/vacancies)
