--- src/routes/about/meta.ts
@@ -1,9 +1,0 @@
-import type { StrictPost } from '$lib/types'
-
-export const meta: StrictPost = {
-	title: 'About Us',
-	description: 'Staff behind PauseAI',
-	date: '2025-11-06',
-	slug: 'about',
-	categories: []
-}
