+++ src/routes/api/posts/+server.ts
@@ -6,7 +6,6 @@ import { meta as quotesMeta } from '../../quotes/meta'
 import { meta as emailBuilderMeta } from '../../email-builder/meta'
 import { meta as statementMeta } from '../../statement/meta'
 import { meta as dearSirDemisMeta } from '../../dear-sir-demis-2025/meta'
-import { meta as aboutMeta } from '../../about/meta'
 import { meta as contactMeta } from '../../contact-us/meta'
 import type { RequestHandler } from './$types'
 
@@ -20,7 +19,6 @@ const hardCodedPages: StrictPost[] = [
 	emailBuilderMeta,
 	statementMeta,
 	dearSirDemisMeta,
-	aboutMeta,
 	contactMeta
 ]
